backend/requirements.txt
fastapi
uvicorn
beanie
motor
pydantic
python-dotenv
python-multipart
PyPDF2
python-docx
pdfplumber
reportlab
scikit-learn
sentence-transformers
google-auth>=2.26.1
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
pymongo
certifi
prometheus-client
httpx
pandas
numpy
sentence-transformers
PyJWT
bcrypt
python-jose[cryptography]
email-validator
google-auth-oauthlib
google-auth
torch
tqdm
requests
redis
aioredis
pika
aio-pika
langchain
langchain-core
langgraph
pdfplumber
langchain-mcp-adapters
mcp
